- initial project structure to split apart cloud and server for xray - add new additional configuration options
